# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DiagnosticSessionResponseRequest": {
        "properties": {
          "correct": {
            "title": "Correct",
            "type": "boolean"
          },
          "item_id": {
            "format": "uuid",
            "title": "Item Id",
            "type": "string"
          },
          "response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          }
        },
        "required": [
          "item_id",
          "correct"
        ],
        "title": "DiagnosticSessionResponseRequest",
        "type": "object"
      },
      "DiagnosticSessionStartRequest": {
        "properties": {
          "caps_ref": {
            "title": "Caps Ref",
            "type": "string"
          },
          "learner_id": {
            "format": "uuid",
            "title": "Learner Id",
            "type": "string"
          },
          "theta": {
            "default": 0.0,
            "title": "Theta",
            "type": "number"
          }
        },
        "required": [
          "learner_id",
          "caps_ref"
        ],
        "title": "DiagnosticSessionStartRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewItemRequest": {
        "properties": {
          "quality_score": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          },
          "review_status": {
            "pattern": "^(draft|ai_generated|human_reviewed|approved|retired)$",
            "title": "Review Status",
            "type": "string"
          }
        },
        "required": [
          "review_status"
        ],
        "title": "ReviewItemRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-powered adaptive learning platform — Grade R to 7. CAPS-aligned. POPIA-compliant.",
    "title": "EduBoost SA V2",
    "version": "1.0.0-rc1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/diagnostics/coverage": {
      "get": {
        "operationId": "get_item_bank_coverage_api_v2_diagnostics_coverage_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Item Bank Coverage",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/api/v2/diagnostics/item-bank/items/{item_id}": {
      "get": {
        "operationId": "get_item_bank_item_api_v2_diagnostics_item_bank_items__item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Item Bank Item",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/api/v2/diagnostics/item-bank/items/{item_id}/review": {
      "post": {
        "operationId": "review_item_bank_item_api_v2_diagnostics_item_bank_items__item_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Item Bank Item",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/api/v2/diagnostics/items/{learner_id}": {
      "get": {
        "operationId": "get_diagnostic_items_api_v2_diagnostics_items__learner_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Diagnostic Items",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/api/v2/diagnostics/sessions": {
      "post": {
        "operationId": "start_diagnostic_session_api_v2_diagnostics_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosticSessionStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Diagnostic Session",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/api/v2/diagnostics/sessions/{session_id}/next-item": {
      "get": {
        "operationId": "diagnostic_next_item_api_v2_diagnostics_sessions__session_id__next_item_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "caps_ref",
            "required": true,
            "schema": {
              "title": "Caps Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Diagnostic Next Item",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/api/v2/diagnostics/sessions/{session_id}/recover": {
      "get": {
        "operationId": "recover_diagnostic_session_api_v2_diagnostics_sessions__session_id__recover_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Recover Diagnostic Session",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/api/v2/diagnostics/sessions/{session_id}/respond": {
      "post": {
        "operationId": "diagnostic_respond_api_v2_diagnostics_sessions__session_id__respond_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosticSessionResponseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Diagnostic Respond",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/api/v2/diagnostics/submit": {
      "post": {
        "operationId": "submit_diagnostic_api_v2_diagnostics_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosticSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Diagnostic",
        "tags": [
          "diagnostics"
        ]
      }
    },
...
    "/v2/diagnostics/coverage": {
      "get": {
        "operationId": "get_item_bank_coverage_v2_diagnostics_coverage_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Item Bank Coverage",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/v2/diagnostics/item-bank/items/{item_id}": {
      "get": {
        "operationId": "get_item_bank_item_v2_diagnostics_item_bank_items__item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Item Bank Item",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/v2/diagnostics/item-bank/items/{item_id}/review": {
      "post": {
        "operationId": "review_item_bank_item_v2_diagnostics_item_bank_items__item_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Item Bank Item",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/v2/diagnostics/items/{learner_id}": {
      "get": {
        "operationId": "get_diagnostic_items_v2_diagnostics_items__learner_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Diagnostic Items",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/v2/diagnostics/sessions": {
      "post": {
        "operationId": "start_diagnostic_session_v2_diagnostics_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosticSessionStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Diagnostic Session",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/v2/diagnostics/sessions/{session_id}/next-item": {
      "get": {
        "operationId": "diagnostic_next_item_v2_diagnostics_sessions__session_id__next_item_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "caps_ref",
            "required": true,
            "schema": {
              "title": "Caps Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Diagnostic Next Item",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/v2/diagnostics/sessions/{session_id}/recover": {
      "get": {
        "operationId": "recover_diagnostic_session_v2_diagnostics_sessions__session_id__recover_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Recover Diagnostic Session",
        "tags": [
          "diagnostics"
        ]
      }
    },
    "/v2/diagnostics/sessions/{session_id}/respond": {
      "post": {
        "operationId": "diagnostic_respond_v2_diagnostics_sessions__session_id__respond_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosticSessionResponseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Diagnostic Respond",
        "tags": [
          "diagnostics"
        ]
      }
    },
...
  },
...
}